semana04/dia3/requirements.txt
flask
python-dotenv
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
mysqlclient